Remove rootshell
