+++ clients/apps/web/src/app/(checkout)/layout.tsx
@@ -1,6 +1,8 @@
-import '@/styles/globals.css'
+import { Metadata } from 'next/types'
 
-import RootShell from '../RootShell'
+export const metadata: Metadata = {
+  robots: { index: false, follow: false },
+}
 
 export default async function CheckoutLayout({
   children,
@@ -8,34 +10,29 @@ export default async function CheckoutLayout({
   children: React.ReactNode
 }) {
   return (
-    <RootShell
-      head={
-        <>
-          <link
-            rel="preload"
-            href="/fonts/Inter-Regular.woff2"
-            as="font"
-            type="font/woff2"
-            crossOrigin=""
-          />
-          <link
-            rel="preload"
-            href="/fonts/Inter-Medium.woff2"
-            as="font"
-            type="font/woff2"
-            crossOrigin=""
-          />
-          <link
-            rel="preload"
-            href="/fonts/InterDisplay-SemiBold.woff2"
-            as="font"
-            type="font/woff2"
-            crossOrigin=""
-          />
-        </>
-      }
-    >
+    <>
+      <link
+        rel="preload"
+        href="/fonts/Inter-Regular.woff2"
+        as="font"
+        type="font/woff2"
+        crossOrigin=""
+      />
+      <link
+        rel="preload"
+        href="/fonts/Inter-Medium.woff2"
+        as="font"
+        type="font/woff2"
+        crossOrigin=""
+      />
+      <link
+        rel="preload"
+        href="/fonts/InterDisplay-SemiBold.woff2"
+        as="font"
+        type="font/woff2"
+        crossOrigin=""
+      />
       {children}
-    </RootShell>
+    </>
   )
 }
