--- clients/apps/web/src/app/RootShell.tsx
@@ -1,69 +1,0 @@
-import SandboxBanner from '@/components/Sandbox/SandboxBanner'
-import { getExperimentNames } from '@/experiments'
-import { getDistinctId } from '@/experiments/distinct-id'
-import { ExperimentProvider } from '@/experiments/ExperimentProvider'
-import { getExperiments } from '@/experiments/server'
-import { UserContextProvider } from '@/providers/auth'
-import { getServerSideAPI } from '@/utils/client/serverside'
-import { CONFIG } from '@/utils/config'
-import { getAuthenticatedUser, getUserOrganizations } from '@/utils/user'
-import { schemas } from '@polar-sh/client'
-import { PHASE_PRODUCTION_BUILD } from 'next/constants'
-import {
-  NavigationHistoryProvider,
-  PolarNuqsProvider,
-  PolarPostHogProvider,
-  PolarQueryClientProvider,
-} from './providers'
-
-export default async function RootShell({
-  head,
-  children,
-}: {
-  head?: React.ReactNode
-  children: React.ReactNode
-}) {
-  const api = await getServerSideAPI()
-
-  let authenticatedUser: schemas['UserRead'] | undefined = undefined
-  let userOrganizations: schemas['Organization'][] = []
-
-  try {
-    authenticatedUser = await getAuthenticatedUser()
-    userOrganizations = await getUserOrganizations(api)
-  } catch (e) {
-    if (process.env.NEXT_PHASE !== PHASE_PRODUCTION_BUILD) {
-      throw e
-    }
-  }
-
-  const distinctId = await getDistinctId()
-  const experimentVariants = await getExperiments(getExperimentNames(), {
-    distinctId,
-  })
-
-  return (
-    <html lang="en" suppressHydrationWarning className="antialiased">
-      <head>{head}</head>
-      <body style={{ textRendering: 'optimizeLegibility' }}>
-        <ExperimentProvider experiments={experimentVariants}>
-          <UserContextProvider
-            user={authenticatedUser}
-            userOrganizations={userOrganizations}
-          >
-            <PolarPostHogProvider distinctId={distinctId}>
-              <PolarQueryClientProvider>
-                <PolarNuqsProvider>
-                  <NavigationHistoryProvider>
-                    {CONFIG.IS_SANDBOX && <SandboxBanner />}
-                    {children}
-                  </NavigationHistoryProvider>
-                </PolarNuqsProvider>
-              </PolarQueryClientProvider>
-            </PolarPostHogProvider>
-          </UserContextProvider>
-        </ExperimentProvider>
-      </body>
-    </html>
-  )
-}
